+++ packages/react/src/components/Furniture/Bands/HorizontalBand/HorizontalBand.mdx
