+++ packages/react/src/components/Furniture/Bands/HorizontalBand/index.js
@@ -1,4 +1,4 @@
-import { withSVG } from "../../../hoc";
+import { withSVG } from "../../../../hoc";
 import { HorizontalBand as HorizontalBandBase } from "./HorizontalBand";
 
 const HorizontalBand = withSVG(HorizontalBandBase, "horizontal-band");
