+++ packages/react/src/components/Furniture/Bands/VerticalBand/VerticalBand.mdx
