+++ packages/react/src/components/Furniture/Bands/VerticalBand/VerticalBand.stories.jsx
@@ -4,21 +4,21 @@ import { useSelector } from "react-redux";
 import PropTypes from "prop-types";
 import { Provider } from "react-redux";
 
-import { HorizontalBand } from ".";
-import mdx from "./HorizontalBand.mdx";
+import { VerticalBand } from ".";
+import mdx from "./VerticalBand.mdx";
 
-import { sales_records_dataset } from "../../../../data/sales_records_dataset";
-import { Scatter } from "../../Plots";
-import { XYChart } from "../../XYChart";
-import { XAxis, YAxis } from "../../Axis";
-import { chartSelectors } from "../../../store";
-import { themes } from "../../../themes";
+import { sales_records_dataset } from "../../../../../data/sales_records_dataset";
+import { Scatter } from "../../../Plots";
+import { XYChart } from "../../../XYChart";
+import { XAxis, YAxis } from "../../../Axis";
+import { chartSelectors } from "../../../../store";
+import { themes } from "../../../../themes";
 
-import { createMockStorybookStore } from "../../../testUtils";
+import { createMockStorybookStore } from "../../../../testUtils";
 
 export default {
-    title: "Components/Bands/HorizontalBand",
-    component: HorizontalBand,
+    title: "Components/Bands/VerticalBand",
+    component: VerticalBand,
     parameters: {
         docs: {
             page: mdx,
@@ -27,7 +27,7 @@ export default {
     },
 };
 
-const HorizontalBandTemplate = () => {
+const VerticalBandTemplate = () => {
     const store = createMockStorybookStore({
         chart: {
             theme: themes.light,
@@ -36,10 +36,10 @@ const HorizontalBandTemplate = () => {
                 height: 200,
             },
             scales: {
-                "Total Profit": d3
+                "Units Sold": d3
                     .scaleLinear()
                     .domain([0, 1000])
-                    .range([200, 0]),
+                    .range([0, 200]),
             },
         },
     });
@@ -47,7 +47,7 @@ const HorizontalBandTemplate = () => {
     return (
         <Provider store={store}>
             <svg width="500" height="200">
-                <HorizontalBand yStop={500} y="Total Profit" fill="steelblue" stroke="red" opacity={0.3} />
+                <VerticalBand xStop={500} x="Units Sold" fill="steelblue" stroke="red" opacity={0.3} />
             </svg>
         </Provider>
     );
@@ -66,16 +66,16 @@ const ScatterWithRectsTemplate = () => {
         >
             <YAxis fields={[y]} />
             <XAxis fields={[x]} />
-            <HorizontalBand y={y} yStop={300000} fill="red" opacity={0.1} />
-            <HorizontalBand y={y} yStart={300000} yStop={900000} fill="orange" opacity={0.1} />
-            <HorizontalBand y={y} yStart={900000} fill="green" opacity={0.1} />
+            <VerticalBand x={x} xStop={2000} fill="red" opacity={0.1} />
+            <VerticalBand x={x} xStart={2000} xStop={7000} fill="orange" opacity={0.1} />
+            <VerticalBand x={x} xStart={7000} fill="green" opacity={0.1} />
             <Scatter x={x} y={y} />
         </XYChart>
     );
 };
 
-export const Default = HorizontalBandTemplate.bind({});
-Default.storyName = "HorizontalBand";
+export const Default = VerticalBandTemplate.bind({});
+Default.storyName = "VerticalBand";
 
 export const ThresholdsExample = ScatterWithRectsTemplate.bind({});
 ThresholdsExample.storyName = "Thresholds Example";
