+++ packages/react/src/components/Furniture/Bands/VerticalBand/index.js
@@ -1,4 +1,4 @@
-import { withSVG } from "../../../hoc";
+import { withSVG } from "../../../../hoc";
 import { VerticalBand as VerticalBandBase } from "./VerticalBand";
 
 const VerticalBand = withSVG(VerticalBandBase, "vertical-band");
