+++ packages/react/src/components/Furniture/Bands/index.js
