+++ packages/react/src/components/Furniture/Polygon/Polygon.jsx
