+++ packages/react/src/components/Furniture/Polygon/Polygon.mdx
