+++ packages/react/src/components/Furniture/Polygon/Polygon.stories.jsx
