+++ packages/react/src/components/Furniture/Polygon/Polygon.unit.jsx
