+++ packages/react/src/components/Furniture/Polygon/index.js
@@ -1,4 +1,4 @@
-import { withSVG } from "../../hoc";
+import { withSVG } from "../../../hoc";
 import { Polygon as PolygonBase } from "./Polygon";
 
 const Polygon = withSVG(PolygonBase, "polygon");
