+++ packages/react/src/components/Furniture/index.js
@@ -1,0 +1,2 @@
+export * from "./Bands";
+export * from "./Polygon";
